+++ hypervisor/include/arch/x86/board.h
@@ -31,7 +31,4 @@ extern struct platform_clos_info platform_mba_clos_array[MAX_PLATFORM_CLOS_NUM];
 extern const struct cpu_state_table board_cpu_state_tbl;
 extern const union pci_bdf plat_hidden_pdevs[MAX_HIDDEN_PDEVS_NUM];
 
-/* board specific functions */
-void create_prelaunched_vm_e820(struct acrn_vm *vm);
-
 #endif /* BOARD_H */
